test/test_pubsub.py: refactor pubsub tests
create a fuction to avoid including duplicated
code to create a pubsub object.
